flyttet pa soketypene
