+++ src/types/soke-typer.ts
@@ -1,0 +1,5 @@
+export type SokeTyper = {
+    veilederNavIdent?: string;
+    bedriftNr?: string;
+    deltakerFnr?: string;
+};
